backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.11.3
pydantic==2.3.0
pydantic-settings==2.0.3
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
redis==5.0.0
httpx==0.25.1
pytest==7.4.2
pytest-asyncio==0.21.1
aioredis==2.0.1
python-dotenv==1.1.1